feat: chat can now be used (limited to certain user)
